Refactor ilastik options
* Use special "MESSAGE" visibility for providing path examples

* Remove value auto-loading from persistent storage because
  it interferes with programmatic options usage

* Simplify field labels wording

* Change class name to be more consistent with built-in ImageJ plugins
